+++ src/app/condition/condition.module.js
@@ -3,6 +3,6 @@
   // 条件模块, 条件容器类, 维度类, 条件业务交互
 
   angular
-    .module('pf.condition', ['pf.dimension', 'pf.sheet']);
+    .module('pf.condition', ['pf.dimension', 'pf.sheet', 'pf.indicator']);
 
 })();