+++ src/app/indicator/indicator.module.js
@@ -1,0 +1,8 @@
+(function() {
+  'use strict';
+  // 指标模块
+
+  angular
+    .module('pf.indicator', []);
+
+})();